pyproject.toml
[project]
name = "mywork-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "playwright>=1.49.1",
    "pydantic-settings[yaml]>=2.7.0",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "surrealdb==0.3.2",
    "xlwings>=0.33.5",
]

[tool.setuptools]
package-dir = {"" = "src"}
